re factored into best practice file structure
