+++ api/server.js
@@ -1,0 +1,13 @@
+const express = require('express');
+const helmet = require('helmet');
+
+const zoosRouter = require('../routers/zoos-router.js');
+
+const server = express();
+
+server.use(helmet());
+server.use(express.json());
+
+server.use('/api/zoos', zoosRouter);
+
+module.exports = server;
